[refactor]: Generator 제거; 거대한 도메인에 대한 불필요한 확장 기능 제거
