--- src/main/java/ladder/domain/Generator.java
@@ -1,5 +1,0 @@
-package ladder.domain;
-
-public interface Generator<T> {
-    T generate();
-}
