+++ src/test/java/ladder/domain/LadderTest.java
@@ -79,4 +79,33 @@ class LadderTest {
                 ))
         );
     }
+
+    @Test
+    @DisplayName("생성자로 만들어진 ladder 의 move 가 정상적인지 확인")
+    void constructor() {
+        int height = 3;
+        int sizeOfPersons = 5;
+        Ladder ladder = new Ladder(sizeOfPersons, height, new DirectionStrategy() {
+            @Override
+            Direction get() {
+                return get(false);
+            }
+        });
+        assertAll(
+                () -> assertThatExceptionOfType(BadPositionException.class)
+                        .isThrownBy(() -> ladder.move(sizeOfPersons)),
+                () -> assertThatExceptionOfType(BadPositionException.class)
+                        .isThrownBy(() -> ladder.move(sizeOfPersons + 7)),
+                () -> assertThat(ladder.move(0))
+                        .isEqualTo(1),
+                () -> assertThat(ladder.move(1))
+                        .isEqualTo(0),
+                () -> assertThat(ladder.move(2))
+                        .isEqualTo(3),
+                () -> assertThat(ladder.move(3))
+                        .isEqualTo(2),
+                () -> assertThat(ladder.move(4))
+                        .isEqualTo(4)
+        );
+    }
 }
